refactor: move all game files under game/ folder
also moved test files to match src files
