+++ backend/src/main/game/room/waypoint/waypoint_b_pojo.py
@@ -1,5 +1,5 @@
 from backend.src.main.game.monster.values import UniqueDungeonCardValues
-from backend.src.main.room.waypoint.waypoint_pojo import WaypointPOJO
+from backend.src.main.game.room.waypoint.waypoint_pojo import WaypointPOJO
 
 
 class WaypointB(WaypointPOJO):
