+++ backend/src/main/game/tile/tile_geometry.py
@@ -1,7 +1,7 @@
 from backend.src.main.game.room.room import AbstractRoomCard
-from backend.src.main.room.waypoint.waypoint_pojo import WaypointPOJO
-from backend.src.main.tile.center_tile import CenterTile
-from backend.src.main.tile.shift_tile import ShiftTile
+from backend.src.main.game.room.waypoint.waypoint_pojo import WaypointPOJO
+from backend.src.main.game.tile.center_tile import CenterTile
+from backend.src.main.game.tile.shift_tile import ShiftTile
 
 
 class TileGeometry:
