+++ backend/src/main/game/tile/tile_geometry_util.py
