+++ backend/src/test/game/dungeon/test_dungeon.py
@@ -2,15 +2,15 @@
 from unittest.mock import MagicMock, call, patch
 
 import pytest
-from backend.src.main.game.room.room import AbstractRoomCard
 
-from backend.src.main.dungeon.dungeon import RandomDungeonGenerator
+from backend.src.main.game.dungeon.dungeon import RandomDungeonGenerator
 from backend.src.main.game.monster.random_monster_card import AbstractMonsterCard
-from backend.src.main.room.constructed_room import ConstructedRoom
-from backend.src.main.room.waypoint.waypoint_a_pojo import WaypointA
-from backend.src.main.tile.tile_geometry import TileGeometry
+from backend.src.main.game.room.constructed_room import ConstructedRoom
+from backend.src.main.game.room.room import AbstractRoomCard
+from backend.src.main.game.room.waypoint.waypoint_a_pojo import WaypointA
+from backend.src.main.game.tile.tile_geometry import TileGeometry
 from backend.src.main.wrappers.random_wrapper import RandomWrapper
-from backend.src.test.dungeon import util
+from backend.src.test.game.dungeon import util
 
 
 @pytest.fixture(name='dungeon_generator')
