+++ backend/src/test/game/monster/test_cutthroat.py
