+++ backend/src/test/game/room/test_constructed_room.py
@@ -3,8 +3,8 @@ import pytest
 from backend.src.main.game.monster.concrete_monster_cards.cutthroat import Cutthroat
 from backend.src.main.game.monster.values import NumberedRoomTileValues, DungeonCardValues
 from backend.src.main.game.room.concrete_room_cards.den import Den
+from backend.src.main.game.room.constructed_room import ConstructedRoom
 from backend.src.main.game.room.room import Tile
-from backend.src.main.room.constructed_room import ConstructedRoom
 
 
 def test_numbered_tile_is_numbered_tile():
