+++ backend/src/test/game/tile/test_center_tile.py
@@ -1,11 +1,12 @@
 import pytest
+
 from backend.src.main.game.monster.values import DungeonCardValues, UniqueDungeonCardValues
 from backend.src.main.game.room.concrete_room_cards.burrow import Burrow
 from backend.src.main.game.room.concrete_room_cards.hovel import Hovel
 from backend.src.main.game.room.concrete_room_cards.tunnel import Tunnel
-from backend.src.main.room.waypoint.waypoint_pojo import WaypointPOJO
-from backend.src.main.tile.center_tile import CenterTile
-from backend.src.main.tile.tile import Tile
+from backend.src.main.game.room.waypoint.waypoint_pojo import WaypointPOJO
+from backend.src.main.game.tile.center_tile import CenterTile
+from backend.src.main.game.tile.tile import Tile
 
 
 @pytest.fixture(name='waypoint_pojo_a', scope='function')
