+++ backend/src/test/game/tile/test_tile.py
@@ -1,5 +1,5 @@
 from backend.src.main.game.monster.values import NumberedRoomTileValues
-from backend.src.main.tile.tile import Tile
+from backend.src.main.game.tile.tile import Tile
 
 
 def test_tile_with_same_attributes_are_equal():
